+++ frontend/src/app/router/routes.ts
@@ -3,14 +3,14 @@ import { RouteRecordRaw } from 'vue-router';
 const routes: RouteRecordRaw[] = [
   {
     path: '/',
-    component: () => import('pages/PHome.vue'),
+    component: () => import('processes/layouts/LHome.vue'),
   },
 
   // Always leave this as last one,
   // but you can also remove it
   {
     path: '/:catchAll(.*)*',
-    component: () => import('pages/POther.vue'),
+    component: () => import('processes/layouts/LOther.vue'),
   },
 ];
