--- frontend/src/pages/PHome.vue
@@ -1,8 +1,0 @@
-<!-- eslint-disable vue/multi-word-component-names -->
-<script setup lang="ts">
-import LHome from 'processes/layouts/LHome.vue';
-</script>
-
-<template>
-  <LHome></LHome>
-</template>
