+++ frontend/src/pages/PLanding.vue
@@ -1,0 +1,14 @@
+<!-- eslint-disable vue/multi-word-component-names -->
+<script setup lang="ts">
+import { WAboutUs } from 'widgets/WAboutUs';
+import { WAdvances } from 'widgets/WAdvances';
+import { WHeader } from 'widgets/WHeader';
+import { WHome } from 'widgets/WHome';
+</script>
+
+<template>
+  <WHeader></WHeader>
+  <WHome />
+  <WAboutUs></WAboutUs>
+  <WAdvances></WAdvances>
+</template>
