+++ frontend/src/processes/layouts/LHome.vue
@@ -1,15 +1,9 @@
 <script setup lang="ts">
-import { WAboutUs } from 'widgets/AboutUs';
-import { WAdvances } from 'widgets/Advances';
-import { WHeader } from 'widgets/Header';
-import { WHome } from 'widgets/Home';
+import PLanding from 'pages/PLanding.vue';
 </script>
 
 <template>
   <q-layout view="hHh lpr fFf">
-    <WHeader></WHeader>
-    <WHome />
-    <WAboutUs></WAboutUs>
-    <WAdvances></WAdvances>
+    <PLanding></PLanding>
   </q-layout>
 </template>
