--- frontend/src/shared/ui/Logo/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as WLogo } from './WLogo.vue';
