+++ frontend/src/shared/ui/SButton/index.ts
