+++ frontend/src/shared/ui/SButton/ui/index.ts
