+++ frontend/src/shared/ui/SLogo/index.ts
