+++ frontend/src/shared/ui/SLogo/ui/SLogo.vue
