+++ frontend/src/shared/ui/SLogo/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as SLogo } from './SLogo.vue';
