+++ frontend/src/shared/ui/SPrettyHeader/index.ts
