+++ frontend/src/shared/ui/SPrettyHeader/ui/SPrettyHeader.vue
