+++ frontend/src/shared/ui/SStructure/index.ts
