+++ frontend/src/shared/ui/SStructure/ui/SStructure.vue
