+++ frontend/src/shared/ui/SStructure/ui/index.ts
