+++ frontend/src/widgets/WAboutUs/index.ts
