+++ frontend/src/widgets/WAboutUs/model/i18n/en.ts
