+++ frontend/src/widgets/WAboutUs/model/i18n/index.ts
@@ -1,7 +1,7 @@
 import enUS from './en';
 import ruRU from './ru';
 
-export const THeader = {
+export const TAboutUsI18n = {
   'en-US': enUS,
   'ru-RU': ruRU,
 };
