+++ frontend/src/widgets/WAboutUs/model/i18n/ru.ts
