+++ frontend/src/widgets/WAboutUs/model/index.ts
