+++ frontend/src/widgets/WAboutUs/ui/index.ts
