+++ frontend/src/widgets/WAdvances/model/i18n/index.ts
@@ -1,7 +1,7 @@
 import enUS from './en';
 import ruRU from './ru';
 
-export const TAdvances = {
+export const TAdvancesI18n = {
   'en-US': enUS,
   'ru-RU': ruRU,
 };
