+++ frontend/src/widgets/WAdvances/model/i18n/ru.ts
