+++ frontend/src/widgets/WAdvances/ui/AdvanceCard.vue
