+++ frontend/src/widgets/WHeader/model/i18n/index.ts
@@ -1,7 +1,7 @@
 import enUS from './en';
 import ruRU from './ru';
 
-export const TAboutUs = {
+export const THeaderI18n = {
   'en-US': enUS,
   'ru-RU': ruRU,
 };
