+++ frontend/src/widgets/WHeader/model/i18n/ru.ts
