+++ frontend/src/widgets/WHome/index.ts
