+++ frontend/src/widgets/WHome/model/index.ts
@@ -1,7 +1,7 @@
 import en from './en';
 import ru from './ru';
 
-export const THome = {
+export const THomeI18n = {
   'en-US': en,
   'ru-RU': ru,
 };
