+++ frontend/src/widgets/WHome/ui/HomeChip.vue
