+++ frontend/src/widgets/WHome/ui/index.ts
